Time: 8 ms (41.38%), Space: 14.1 MB (78.83%) - LeetHub

+++ 1482-how-many-numbers-are-smaller-than-the-current-number/1482-how-many-numbers-are-smaller-than-the-current-number.cpp
@@ -1,12 +1,12 @@
 class Solution {
 public:
     vector<int> smallerNumbersThanCurrent(vector<int>& nums) {
-        unsigned short numsSize = nums.size();
+        int numsSize = nums.size();
         vector<int> answer(numsSize);
 
-        for (unsigned short i = 0; i < numsSize - 1; ++i)
+        for (int i = 0; i < numsSize - 1; ++i)
         {
-            for (unsigned short j = i + 1; j < numsSize; ++j)
+            for (int j = i + 1; j < numsSize; ++j)
             {
                 if (nums[i] > nums[j])
                     ++answer[i];
